Moved `combo` function to a distinct helper crate Rust crates "registers_view", "memory_view" now list their dependencies in "Depends" section, not in "Source"
