+++ src/helpers/combo/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "combo"
+version = "0.1.0"
+authors = ["slnpacifist <slnpacifist@gmail.com>"]
+
+[dependencies]
+prodbg_api = { path = "../../../api/rust/prodbg" }
