+++ units.prodbg.lua
@@ -139,6 +139,28 @@ RustCrate  {
 
 -----------------------------------------------------------------------------------------------------------------------
 
+RustCrate  {
+	Name = "combo",
+	CargoConfig = "src/helpers/combo/Cargo.toml",
+	Sources = {
+		get_rs_src("src/helpers/combo"),
+	},
+	Depends = { "prodbg_api" }
+}
+
+-----------------------------------------------------------------------------------------------------------------------
+
+RustCrate  {
+	Name = "char_editor",
+	CargoConfig = "src/helpers/char_editor/Cargo.toml",
+	Sources = {
+		get_rs_src("src/helpers/char_editor"),
+	},
+	Depends = { "prodbg_api" }
+}
+
+-----------------------------------------------------------------------------------------------------------------------
+
 --[[
 
 RustProgram {
@@ -171,7 +193,7 @@ RustProgram {
     Depends = { "lua", "remote_api", "stb", "bgfx_native", "bgfx", "ui",
     			"imgui", "tinyxml2", "capstone", "renderer", "scintilla",
     			"imgui_sys", "core", "viewdock", "settings", "prodbg_api",
-    			"settings", "project", "serde_macros", "number_view" },
+    			"settings", "project", "serde_macros", "number_view", "char_editor", "combo" },
 }
 
 -----------------------------------------------------------------------------------------------------------------------
